Fix calculation of the marginal probability of the slim path.
It was done in integer domain before, which doesn't make sense when
the quantities are probabilities.

+++ ontologizer/src/ontologizer/calculation/b2g/Bayes2GOCalculation.java
@@ -559,7 +559,7 @@ public class Bayes2GOCalculation implements ICalculation, ISlimCalculation, IPro
 			{
 				for (int t = 0; t < numTerms; t++)
 				{
-					res[t] = fixedAlphaBetaScore.termActivationCounts[t] / fixedAlphaBetaScore.numRecords;
+					res[t] = (double)fixedAlphaBetaScore.termActivationCounts[t] / fixedAlphaBetaScore.numRecords;
 				}
 			}
